src/python/requirements.txt
python-igraph=0.8.2=pypi_0
